requirements.txt.txt
streamlit
wikipedia-api
google-generativeai